chatbot_backend/requirements.txt
pandas>=1.5.0
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.10.0
sentence-transformers>=2.2.0
torch>=1.13.0
transformers>=4.20.0
accelerate>=0.20.0
peft>=0.4.0
bitsandbytes>=0.41.0
faiss-cpu>=1.7.0
google-generativeai==0.8.3
python-dotenv==1.0.1
colorama==0.4.6
rich==13.7.0
typer==0.12.0
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0
requests>=2.28.0 
